+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Support for `slug_id` URL param in `WordpressPostContainer`, `WordpressNavigation`, and `WordpressPage` components
+
 ## [2.12.0] - 2021-07-26
 
 ### Added